some components moving and renaming
